+++ staging/CinderAshSplit/index.ts
@@ -1,0 +1,1 @@
+export { CinderAshSplit, type CinderAshSplitProps } from './CinderAshSplit'