+++ staging/NebulaCoronaHero/index.ts
@@ -1,0 +1,1 @@
+export { NebulaCoronaHero, type NebulaCoronaHeroProps } from './NebulaCoronaHero'